# openapi.yaml
openapi: 3.0.2
info:
  version: 1.0.0
  title: payments-configuration-management-service OpenAPI Documentation
  x-gateway-info:
    name: payments-configuration-management-service
...
paths:
...
  /api/providers/{code}:
    patch:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider
      summary: "Partially update provider by code"
      description: "This endpoint allows you to apply partial updates to a provider by specifying its unique code. It updates the provider and returns the modified entity, reflecting only the changes provided in the request."
      operationId: "updateProvider"
      x-eov-operation-handler: "api/methods"
      parameters:
        - in: path
          name: code
          required: true
          schema:
            type: string
      requestBody:
        content:
          application/json:
            schema:
              type: object
              required: ['isEnabled']
              additionalProperties: false
              properties:
                isEnabled:
                  type: boolean
                  example: true
      responses:
        200:
          description: 'Successful Request'
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Provider"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
  /api/providers/{code}/stp-rules:
    get:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider
      summary: "Retrieve stp rules configuration for a particular provider"
      operationId: "getStpProviderRules"
      x-eov-operation-handler: "api/methods"
      description: |
        Retrieves the configuration of Straight-Through Processing (STP) rules for a specific provider identified by the `code` parameter. STP rules define the automated processing logic for handling transactions without manual intervention.
      parameters:
        - in: path
          name: code
          required: true
          description: The unique identifier of the provider.
          schema:
            type: string
            example: "bankwire"
      responses:
        200:
          description: 'Successful Request'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StpProviderRulesDto'
        400:
          $ref: "#/components/responses/BadRequest"
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        500:
          $ref: "#/components/responses/ServerError"
    put:
      security:
        - authorization: [ ]
        - apiKey: [ ]
      tags:
        - Provider
      summary: "Update the STP rules configuration for a particular provider"
      operationId: "updateStpProviderRules"
      x-eov-operation-handler: "api/methods"
      description: |
        Updates the configuration of Straight-Through Processing (STP) rules for a specific provider identified by the code parameter. STP rules define the automated processing logic for handling transactions without manual intervention.
      parameters:
        - in: path
          name: code
          required: true
          description: The unique identifier of the provider.
          schema:
            type: string
            example: "bankwire"
      requestBody:
        content:
          application/json:
            schema:
              type: array
              items:
                $ref: '#/components/schemas/StpProviderRulesDto'
      responses:
        200:
          description: 'Successful Request'
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: '#/components/schemas/StpProviderRulesDto'
        401:
          $ref: "#/components/responses/Unauthorized"
        403:
          $ref: "#/components/responses/Forbidden"
        404:
          $ref: "#/components/responses/NotFound"
        409:
          $ref: "#/components/responses/Conflict"
        500:
          $ref: "#/components/responses/ServerError"
...
components:
  securitySchemes:
    authorization:
      type: http
      scheme: bearer
      bearerFormat: JWT
    apiKey:
      type: apiKey
      in: header
      name: x-api-key
...
  schemas:
    Error:
      type: object
      properties:
        code:
          type: string
        message:
          type: string
        requestId:
          type: string
        meta:
          type: object
      required:
        - code
        - requestId
    Authority:
      type: string
      minLength: 1
      maxLength: 20
      example: 'FSCM'
...
    ProviderBase:
      type: object
      properties:
        code:
          type: string
        name:
          type: string
        isEnabled:
          type: boolean
    Provider:
      allOf:
        - $ref: '#/components/schemas/ProviderBase'
        - type: object
          properties:
            type:
              type: string
...
    StpProviderRulesDto:
      type: object
      required: ['countriesAuthorities', 'stpRules', 'isEnabled']
      additionalProperties: false
      properties:
        isEnabled:
          type: boolean
        countriesAuthorities:
          type: array
          minItems: 1
          items:
            type: object
            required: [ 'authority' ]
            properties:
              authority:
                $ref: '#/components/schemas/Authority'
        stpRules:
          type: array
          minItems: 1
          items:
            type: object
            required: [ 'key', 'isEnabled' ]
            properties:
              key:
                type: string
              isEnabled:
                type: boolean
              type:
                type: string
                nullable: true
                enum:
                  - number
                  - list
              value:
                anyOf:
                  - type: string
                    minLength: 1
                    nullable: false
                  - type: array
                    items:
                      type: string
                      minLength: 1
                      nullable: false
                    nullable: true
...
  responses:
    BadRequest:
      description: Bad request
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    NotFound:
      description: Not found
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    Conflict:
      description: Conflict
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Unauthorized:
      description: Unauthorized
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
    Forbidden:
      description: Forbidden
      content:
        application/json:
          schema:
            $ref: "#/components/schemas/Error"
...
    ServerError:
      description: Server error.